refactor(discord-sonia-config): split into multiple services
