--- src/features/discord/users/constants/discord-sonia-config.ts
@@ -1,12 +1,0 @@
-import { IDiscordSoniaConfig } from "../../interfaces/discord-sonia-config";
-
-export const DISCORD_SONIA_CONFIG: IDiscordSoniaConfig = {
-  corporationImageUrl: `https://i.ibb.co/XSB6Vng/icons8-girl-1024.png`,
-  corporationMessageEmbedAuthor: {
-    iconURL: `https://i.ibb.co/XSB6Vng/icons8-girl-1024.png`,
-    name: `Sonia`,
-    url: `https://github.com/Sonia-corporation?type=source`,
-  },
-  id: `unknown`,
-  secretToken: `unknown`,
-};
