+++ src/features/discord/users/enums/discord-sonia-config-value-name.enum.ts
@@ -1,0 +1,8 @@
+export enum DiscordSoniaConfigValueNameEnum {
+  CORPORATION_IMAGE_URL = `corporation image url`,
+  CORPORATION_MESSAGE_EMBED_AUTHOR_ICON_URL = `corporation message embed author icon url`,
+  CORPORATION_MESSAGE_EMBED_AUTHOR_NAME = `corporation message embed author name`,
+  CORPORATION_MESSAGE_EMBED_AUTHOR_URL = `corporation message embed author url`,
+  ID = `id`,
+  SECRET_TOKEN = `secret token`,
+}
