+++ src/features/discord/users/services/config/discord-sonia-config-mutator-service.spec.ts
@@ -1,63 +1,19 @@
-import { IConfigUpdateString } from "../../../config/interfaces/config-update-string";
-import { ConfigService } from "../../../config/services/config-service";
-import { IDiscordSoniaConfig } from "../../interfaces/discord-sonia-config";
-import { IDiscordSoniaCorporationMessageEmbedAuthorConfig } from "../../interfaces/discord-sonia-corporation-message-embed-author-config";
-import { DISCORD_SONIA_CONFIG } from "../constants/discord-sonia-config";
-import { DiscordSoniaConfigService } from "./discord-sonia-config-service";
+import { IConfigUpdateString } from "../../../../config/interfaces/config-update-string";
+import { ConfigService } from "../../../../config/services/config-service";
+import { DiscordSoniaConfigCoreService } from "./discord-sonia-config-core-service";
+import { DiscordSoniaConfigMutatorService } from "./discord-sonia-config-mutator-service";
 
-jest.mock(`../../../config/services/config-service`);
+jest.mock(`../../../../config/services/config-service`);
 
-describe(`DiscordSoniaConfigService`, (): void => {
-  let service: DiscordSoniaConfigService;
+describe(`DiscordSoniaConfigMutatorService`, (): void => {
+  let service: DiscordSoniaConfigMutatorService;
   let configService: ConfigService;
+  let discordSoniaConfigCoreService: DiscordSoniaConfigCoreService;
 
   beforeEach((): void => {
-    service = DiscordSoniaConfigService.getInstance();
+    service = DiscordSoniaConfigMutatorService.getInstance();
     configService = ConfigService.getInstance();
-  });
-
-  describe(`getSonia()`, (): void => {
-    beforeEach((): void => {
-      DISCORD_SONIA_CONFIG.corporationImageUrl = `dummy-corporation-image-url`;
-      DISCORD_SONIA_CONFIG.corporationMessageEmbedAuthor = {
-        iconURL: `dummy-icon-url`,
-        name: `dummy-name`,
-        url: `dummy-url`,
-      };
-      DISCORD_SONIA_CONFIG.id = `dummy-id`;
-      DISCORD_SONIA_CONFIG.secretToken = `dummy-secret-token`;
-    });
-
-    it(`should return the Discord Sonia config`, (): void => {
-      expect.assertions(1);
-
-      const result = service.getSonia();
-
-      expect(result).toStrictEqual({
-        corporationImageUrl: `dummy-corporation-image-url`,
-        corporationMessageEmbedAuthor: {
-          iconURL: `dummy-icon-url`,
-          name: `dummy-name`,
-          url: `dummy-url`,
-        },
-        id: `dummy-id`,
-        secretToken: `dummy-secret-token`,
-      } as IDiscordSoniaConfig);
-    });
-  });
-
-  describe(`getCorporationImageUrl()`, (): void => {
-    beforeEach((): void => {
-      DISCORD_SONIA_CONFIG.corporationImageUrl = `dummy-corporation-image-url`;
-    });
-
-    it(`should return the Discord Sonia config corporation image url`, (): void => {
-      expect.assertions(1);
-
-      const result = service.getCorporationImageUrl();
-
-      expect(result).toStrictEqual(`dummy-corporation-image-url`);
-    });
+    discordSoniaConfigCoreService = DiscordSoniaConfigCoreService.getInstance();
   });
 
   describe(`updateCorporationImageUrl()`, (): void => {
@@ -67,7 +23,7 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
     beforeEach((): void => {
       corporationImageUrl = `dummy-corporation-image-url`;
-      DISCORD_SONIA_CONFIG.corporationImageUrl = `corporation-image-url`;
+      discordSoniaConfigCoreService.corporationImageUrl = `corporation-image-url`;
 
       configServiceGetUpdatedStringSpy = jest
         .spyOn(configService, `getUpdatedString`)
@@ -81,7 +37,7 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
       expect(configServiceGetUpdatedStringSpy).toHaveBeenCalledTimes(1);
       expect(configServiceGetUpdatedStringSpy).toHaveBeenCalledWith({
-        context: `DiscordSoniaConfigService`,
+        context: `DiscordSoniaConfigMutatorService`,
         newValue: `dummy-corporation-image-url`,
         oldValue: `corporation-image-url`,
         valueName: `corporation image url`,
@@ -93,45 +49,9 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
       service.updateCorporationImageUrl(corporationImageUrl);
 
-      expect(DISCORD_SONIA_CONFIG.corporationImageUrl).toStrictEqual(
+      expect(discordSoniaConfigCoreService.corporationImageUrl).toStrictEqual(
         `dummy-corporation-image-url`
       );
-    });
-  });
-
-  describe(`getCorporationMessageEmbedAuthor()`, (): void => {
-    beforeEach((): void => {
-      DISCORD_SONIA_CONFIG.corporationMessageEmbedAuthor = {
-        iconURL: `dummy-icon-url`,
-        name: `dummy-name`,
-        url: `dummy-url`,
-      };
-    });
-
-    it(`should return the Discord Sonia config corporation message embed author`, (): void => {
-      expect.assertions(1);
-
-      const result = service.getCorporationMessageEmbedAuthor();
-
-      expect(result).toStrictEqual({
-        iconURL: `dummy-icon-url`,
-        name: `dummy-name`,
-        url: `dummy-url`,
-      } as IDiscordSoniaCorporationMessageEmbedAuthorConfig);
-    });
-  });
-
-  describe(`getCorporationMessageEmbedAuthorIconUrl()`, (): void => {
-    beforeEach((): void => {
-      DISCORD_SONIA_CONFIG.corporationMessageEmbedAuthor.iconURL = `dummy-icon-url`;
-    });
-
-    it(`should return the Discord Sonia config corporation message embed author icon url`, (): void => {
-      expect.assertions(1);
-
-      const result = service.getCorporationMessageEmbedAuthorIconUrl();
-
-      expect(result).toStrictEqual(`dummy-icon-url`);
     });
   });
 
@@ -142,7 +62,7 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
     beforeEach((): void => {
       iconUrl = `dummy-icon-url`;
-      DISCORD_SONIA_CONFIG.corporationMessageEmbedAuthor.iconURL = `icon-url`;
+      discordSoniaConfigCoreService.corporationMessageEmbedAuthor.iconURL = `icon-url`;
 
       configServiceGetUpdatedStringSpy = jest
         .spyOn(configService, `getUpdatedString`)
@@ -156,7 +76,7 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
       expect(configServiceGetUpdatedStringSpy).toHaveBeenCalledTimes(1);
       expect(configServiceGetUpdatedStringSpy).toHaveBeenCalledWith({
-        context: `DiscordSoniaConfigService`,
+        context: `DiscordSoniaConfigMutatorService`,
         newValue: `dummy-icon-url`,
         oldValue: `icon-url`,
         valueName: `corporation message embed author icon url`,
@@ -169,22 +89,8 @@ describe(`DiscordSoniaConfigService`, (): void => {
       service.updateCorporationMessageEmbedAuthorIconUrl(iconUrl);
 
       expect(
-        DISCORD_SONIA_CONFIG.corporationMessageEmbedAuthor.iconURL
+        discordSoniaConfigCoreService.corporationMessageEmbedAuthor.iconURL
       ).toStrictEqual(`dummy-icon-url`);
-    });
-  });
-
-  describe(`getCorporationMessageEmbedAuthorName()`, (): void => {
-    beforeEach((): void => {
-      DISCORD_SONIA_CONFIG.corporationMessageEmbedAuthor.name = `dummy-name`;
-    });
-
-    it(`should return the Discord Sonia config corporation message embed author name`, (): void => {
-      expect.assertions(1);
-
-      const result = service.getCorporationMessageEmbedAuthorName();
-
-      expect(result).toStrictEqual(`dummy-name`);
     });
   });
 
@@ -195,7 +101,7 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
     beforeEach((): void => {
       name = `dummy-name`;
-      DISCORD_SONIA_CONFIG.corporationMessageEmbedAuthor.name = `name`;
+      discordSoniaConfigCoreService.corporationMessageEmbedAuthor.name = `name`;
 
       configServiceGetUpdatedStringSpy = jest
         .spyOn(configService, `getUpdatedString`)
@@ -209,7 +115,7 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
       expect(configServiceGetUpdatedStringSpy).toHaveBeenCalledTimes(1);
       expect(configServiceGetUpdatedStringSpy).toHaveBeenCalledWith({
-        context: `DiscordSoniaConfigService`,
+        context: `DiscordSoniaConfigMutatorService`,
         newValue: `dummy-name`,
         oldValue: `name`,
         valueName: `corporation message embed author name`,
@@ -222,22 +128,8 @@ describe(`DiscordSoniaConfigService`, (): void => {
       service.updateCorporationMessageEmbedAuthorName(name);
 
       expect(
-        DISCORD_SONIA_CONFIG.corporationMessageEmbedAuthor.name
+        discordSoniaConfigCoreService.corporationMessageEmbedAuthor.name
       ).toStrictEqual(`dummy-name`);
-    });
-  });
-
-  describe(`getCorporationMessageEmbedAuthorUrl()`, (): void => {
-    beforeEach((): void => {
-      DISCORD_SONIA_CONFIG.corporationMessageEmbedAuthor.url = `dummy-url`;
-    });
-
-    it(`should return the Discord Sonia config corporation message embed author url`, (): void => {
-      expect.assertions(1);
-
-      const result = service.getCorporationMessageEmbedAuthorUrl();
-
-      expect(result).toStrictEqual(`dummy-url`);
     });
   });
 
@@ -248,7 +140,7 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
     beforeEach((): void => {
       url = `dummy-url`;
-      DISCORD_SONIA_CONFIG.corporationMessageEmbedAuthor.url = `url`;
+      discordSoniaConfigCoreService.corporationMessageEmbedAuthor.url = `url`;
 
       configServiceGetUpdatedStringSpy = jest
         .spyOn(configService, `getUpdatedString`)
@@ -262,7 +154,7 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
       expect(configServiceGetUpdatedStringSpy).toHaveBeenCalledTimes(1);
       expect(configServiceGetUpdatedStringSpy).toHaveBeenCalledWith({
-        context: `DiscordSoniaConfigService`,
+        context: `DiscordSoniaConfigMutatorService`,
         newValue: `dummy-url`,
         oldValue: `url`,
         valueName: `corporation message embed author url`,
@@ -275,22 +167,8 @@ describe(`DiscordSoniaConfigService`, (): void => {
       service.updateCorporationMessageEmbedAuthorUrl(url);
 
       expect(
-        DISCORD_SONIA_CONFIG.corporationMessageEmbedAuthor.url
+        discordSoniaConfigCoreService.corporationMessageEmbedAuthor.url
       ).toStrictEqual(`dummy-url`);
-    });
-  });
-
-  describe(`getId()`, (): void => {
-    beforeEach((): void => {
-      DISCORD_SONIA_CONFIG.id = `dummy-id`;
-    });
-
-    it(`should return the Discord Sonia config id`, (): void => {
-      expect.assertions(1);
-
-      const result = service.getId();
-
-      expect(result).toStrictEqual(`dummy-id`);
     });
   });
 
@@ -301,7 +179,7 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
     beforeEach((): void => {
       id = `dummy-id`;
-      DISCORD_SONIA_CONFIG.id = `id`;
+      discordSoniaConfigCoreService.id = `id`;
 
       configServiceGetUpdatedStringSpy = jest
         .spyOn(configService, `getUpdatedString`)
@@ -315,7 +193,7 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
       expect(configServiceGetUpdatedStringSpy).toHaveBeenCalledTimes(1);
       expect(configServiceGetUpdatedStringSpy).toHaveBeenCalledWith({
-        context: `DiscordSoniaConfigService`,
+        context: `DiscordSoniaConfigMutatorService`,
         isValueHidden: true,
         newValue: `dummy-id`,
         oldValue: `id`,
@@ -328,21 +206,7 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
       service.updateId(id);
 
-      expect(DISCORD_SONIA_CONFIG.id).toStrictEqual(`dummy-id`);
-    });
-  });
-
-  describe(`getSecretToken()`, (): void => {
-    beforeEach((): void => {
-      DISCORD_SONIA_CONFIG.secretToken = `dummy-secret-token`;
-    });
-
-    it(`should return the Discord Sonia config secret token`, (): void => {
-      expect.assertions(1);
-
-      const result = service.getSecretToken();
-
-      expect(result).toStrictEqual(`dummy-secret-token`);
+      expect(discordSoniaConfigCoreService.id).toStrictEqual(`dummy-id`);
     });
   });
 
@@ -353,7 +217,7 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
     beforeEach((): void => {
       secretToken = `dummy-secret-token`;
-      DISCORD_SONIA_CONFIG.secretToken = `secret-token`;
+      discordSoniaConfigCoreService.secretToken = `secret-token`;
 
       configServiceGetUpdatedStringSpy = jest
         .spyOn(configService, `getUpdatedString`)
@@ -367,7 +231,7 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
       expect(configServiceGetUpdatedStringSpy).toHaveBeenCalledTimes(1);
       expect(configServiceGetUpdatedStringSpy).toHaveBeenCalledWith({
-        context: `DiscordSoniaConfigService`,
+        context: `DiscordSoniaConfigMutatorService`,
         isValueHidden: true,
         newValue: `dummy-secret-token`,
         oldValue: `secret-token`,
@@ -380,7 +244,7 @@ describe(`DiscordSoniaConfigService`, (): void => {
 
       service.updateSecretToken(secretToken);
 
-      expect(DISCORD_SONIA_CONFIG.secretToken).toStrictEqual(
+      expect(discordSoniaConfigCoreService.secretToken).toStrictEqual(
         `dummy-secret-token`
       );
     });
