+++ src/features/discord/users/services/discord-sonia-service.ts
@@ -5,7 +5,7 @@ import { DiscordClientService } from "../../services/discord-client-service";
 import { DiscordSoniaMentalStateEnum } from "../enums/discord-sonia-mental-state.enum";
 import { isDiscordClientUser } from "../functions/is-discord-client-user";
 import { Sonia } from "../types/sonia";
-import { DiscordSoniaConfigService } from "./discord-sonia-config-service";
+import { DiscordSoniaConfigService } from "./config/discord-sonia-config-service";
 
 export class DiscordSoniaService {
   private static _instance: DiscordSoniaService;
